refactor(web): rename query atom helpers
